refactor: make improvements to the db module
- add logging
- rename the module class to DbModule
