+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { DrizzleModule } from './db/db.module';
+import { DbModule } from './db/db.module';
 import { ConfigModule } from '@nestjs/config';
 import { UsersController } from './users/users.controller';
 import { UsersModule } from './users/users.module';
@@ -10,7 +10,7 @@ import { drizzleConfig } from 'drizzle.config';
 
 @Module({
   imports: [
-    DrizzleModule,
+    DbModule,
     // @see: https://docs.nestjs.com/techniques/configuration#configuration for more info
     ConfigModule.forRoot({
       isGlobal: true,
