+++ src/db/db.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { DbConnectionService } from './drizzle.service';
+import { DbConnectionService } from './db.service';
 
 @Module({
   providers: [DbConnectionService],
   exports: [DbConnectionService],
 })
-export class DrizzleModule {}
+export class DbModule {}
