+++ src/users/users.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
-import { DrizzleModule } from 'src/db/db.module';
+import { DbModule } from 'src/db/db.module';
 
 @Module({
   controllers: [UsersController],
   providers: [UsersService],
-  imports: [DrizzleModule],
+  imports: [DbModule],
   exports: [UsersService],
 })
 export class UsersModule {}
